Main

+++ src/others/llvm-link/main.c
@@ -4,7 +4,7 @@ int main() {
    // printf() displays the string inside quotation
    int a = 5;
    int b = 7;
-   int c = adder(a, b);
+   unsigned short c = adder(a, b);
    printf("output is %d\n", c);
    return 0;
 }
